Fixed bug from eylzo and some cleaning

+++ trunk/src/KnownFile.h
@@ -95,7 +95,7 @@ public:
 	uint32	GetAllTimeAccepts() const		{return alltimeaccepted;}
 	void	SetAllTimeAccepts(uint32 new_value) { alltimeaccepted = new_value; };	
 	uint64	GetAllTimeTransfered() const	{return alltimetransferred;}
-	void	SetAllTimeTransfered(uint32 new_value) { alltimetransferred = new_value; };
+	void	SetAllTimeTransfered(uint64 new_value) { alltimetransferred = new_value; };
 	CKnownFile* fileParent;
 	
 private:
